Extract subdivision methods to separate module
