+++ lib/countries/country.rb
@@ -5,6 +5,7 @@ module ISO3166
     extend CountryClassMethods
     extend CountryFinderMethods
     include Emoji
+    include CountrySubdivisionMethods
     attr_reader :data
 
     ISO3166::DEFAULT_COUNTRY_HASH.each do |method_name, _type|
@@ -48,64 +49,6 @@ module ISO3166
 
     def <=>(other)
       to_s <=> other.to_s
-    end
-
-    #  +true+ if this Country has any Subdivisions.
-    def subdivisions?
-      !subdivisions.empty?
-    end
-
-    # @return [Array<ISO3166::Subdivision>] the list of subdivisions for this Country.
-    def subdivisions
-      @subdivisions ||= if data['subdivisions']
-                          ISO3166::Data.create_subdivisions(data['subdivisions'])
-                        else
-                          ISO3166::Data.subdivisions(alpha2)
-                        end
-    end
-
-    # @param types [Array<String>] The locale to use for translations.
-    # @return [Array<ISO3166::Subdivision>] the list of subdivisions of the given type(s) for this Country.
-    def subdivisions_of_types(types)
-      subdivisions.select { |_k, v| types.include?(v.type) }
-    end
-
-    # @return [Array<String>] the list of subdivision types for this country
-    def subdivision_types
-      subdivisions.map { |_k, v| v['type'] }.uniq
-    end
-
-    # @return [Array<String>] the list of humanized subdivision types for this country. Uses ActiveSupport's `#humanize` if available
-    def humanized_subdivision_types
-      if String.instance_methods.include?(:humanize)
-        subdivisions.map { |_k, v| v['type'].humanize }.uniq
-      else
-        subdivisions.map { |_k, v| humanize_string(v['type']) }.uniq
-      end
-    end
-
-    # @param locale [String] The locale to use for translations.
-    # @return [Array<Array>] This Country's subdivision pairs of names and codes.
-    def subdivision_names_with_codes(locale = 'en')
-      subdivisions.map { |k, v| [v.translations[locale] || v.name, k] }
-    end
-
-    # @param locale [String] The locale to use for translations.
-    # @return [Array<String>] A list of subdivision names for this country.
-    def subdivision_names(locale = 'en')
-      subdivisions.map { |_k, v| v.translations[locale] || v.name }
-    end
-
-    def states
-      if RUBY_VERSION =~ /^3\.\d\.\d/
-        warn 'DEPRECATION WARNING: The Country#states method has been deprecated and will be removed in 6.0. Please use Country#subdivisions instead.',
-             uplevel: 1, category: :deprecated
-      else
-        warn 'DEPRECATION WARNING: The Country#states method has been deprecated and will be removed in 6.0. Please use Country#subdivisions instead.',
-             uplevel: 1
-      end
-
-      subdivisions
     end
 
     # +true+ if this country is a member of the European Union.
@@ -170,14 +113,6 @@ module ISO3166
     # @return [String] The name for this Country, in this Country's locale.
     def local_name
       @local_name ||= local_names.first
-    end
-
-    # @param subdivision_str [String] A subdivision name or code to search for. Search includes translated subdivision names.
-    # @return [Subdivision] The first subdivision matching the provided string
-    def find_subdivision_by_name(subdivision_str)
-      subdivisions.select do |k,v|
-        subdivision_str == k || v.name == subdivision_str || v.translations.values.include?(subdivision_str)
-      end.values.first
     end
 
     # @!attribute alpha2
