+++ lib/countries/iso3166.rb
@@ -9,6 +9,7 @@ require 'countries/data/subdivision_methods'
 require 'countries/data'
 require 'countries/structure'
 require 'countries/translations'
+require 'countries/country/country_subdivision_methods'
 require 'countries/country/class_methods'
 require 'countries/country/finder_methods'
 require 'countries/country/emoji'
